move and rename figma sessions
